requirements.txt
pyrtlsdr==0.3.0
numpy>=2.1,<3.0
matplotlib>=3.7.0
setuptools>=70
wheel>=0.43
